pyproject.toml
[project]
name = "camera-driver-python"  
version = "0.1.2"  
description = "A simplified multi camera driver interface"  
readme = "README.md" 
requires-python = ">=3.10"
license = {file = "LICENSE"}

maintainers = [
  {name = "Oliver Batchelor", email = "oliver.batchelor@canterbury.ac.nz" } 
]

dependencies = [
  "beartype",
  "python-dispatch",
  "nvjpeg-torch>=0.1.2",
  "taichi_image",
  "disable_gc",
  "fuzzywuzzy",
  "numpy < 2.0",
  "omegaconf",
  "python-Levenshtein",
]


[tool.setuptools.packages.find]
include = ["camera_driver"]

[project.urls] 
"Homepage" = "https://github.com/uc-vision/camera-driver-python"

[build-system]
requires = ["setuptools>=43.0.0",  "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"


[options]
dependency_links = ["https://pypi.taichi.graphics/simple/"]

[project.scripts]  # Optional
capture_images = "camera_driver.scripts.capture_images:main"
bench_processing = "camera_driver.scripts.bench_processing:main"

[tool.setuptools.package-data]

[tool.pyright]

[tool.ruff]
indent-width = 2


